Added a new feature and some comments
Some comments have been added to various functions to explain them. A
new feature has been added to let the sail out if the roll is greater
than the max desired roll (which can be changed in RoboSail.h). The sail
position is increased to let the sail out. The added number is
determined by the absolute value of the roll (which can be negative if
rolled CCW) multiplied by the compensationFactor (which can be changed
in RoboSail.h).